requirements.txt
python-telegram-bot==21.6
httpx==0.28.1
fastapi==0.115.5
uvicorn==0.32.0